Create README.md
Here's the simple description of your bot in English (A2 level):

What Can This Bot Do?
This is a simple Telegram bot for anime. It helps you find anime fast.

What it does:
Search for Anime: You can type an anime name, and the bot will find information about it.

Random Anime: The bot can show you a random anime if you want something new.

Top Anime: You can see lists of the best anime:

Top overall

Most popular

Top in this season

Top for this year

The bot also shows pictures of anime.

The bot speaks Ukrainian, English, and Danish. You can choose your language.

Why it's good:
The bot is easy to use.

It helps you find anime fast directly in Telegram.

It uses data from the internet to be up-to-date.

We want to add more cool things in the future!